% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{add_point_to_test_ep}
\alias{add_point_to_test_ep}
\title{Add a point to EP}
\usage{
add_point_to_test_ep(
  endpoint,
  x = rnorm(mean = 53.317749, sd = 1, n = 1),
  y = rnorm(mean = -1.0546875, sd = 1, n = 1),
  attributes = list(id = sample(c(1:1000), 1))
)
}
\arguments{
\item{endpoint}{the points endpoint to test against}

\item{x}{the x coordinate of the point (EPSG:4326)}

\item{y}{the y coordinate of the point (EPSG:4326)}

\item{attributes}{the attributes of the point (a named list)}

\item{id}{an integar to give the point an ID (not unique)}
}
\value{
the status code of the response (integar)
}
\description{
Add a test point to the points endpoint
}
\details{
This function is used for testing layer caching & demonstrates how to add features to an arc gis layer
}
