% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{ask_user}
\alias{ask_user}
\title{Ask User}
\usage{
ask_user(prompt)
}
\arguments{
\item{prompt}{a string defining the prompt to ask the user}
}
\value{
TRUE or FALSE depending user input
}
\description{
Ask a user a prompt
}
\details{
This function asks a user a prompt and then parses the response to TRUE/FALSE. It will fail if the user doesn't respond with y/n
}
