% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{assert_that}
\alias{assert_that}
\title{Assert That}
\usage{
assert_that(expr, message = NULL)
}
\arguments{
\item{expr}{an R expression}

\item{message}{the message to print if \code{expr} evaluates to \code{FALSE}. If NULL, the expression is printed like in stopifnot}
}
\description{
Fail and inform
}
