% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cache_method}
\alias{cache_method}
\title{Caching Method}
\usage{
cache_method(layer_details)
}
\arguments{
\item{layer_details}{the layer details object returned be get_layer_details.}
}
\value{
either "layer_edit" for tracking edits to the layer as a whole, "feature_edits" for tracking edits to individual features or NULL for no caching due to lack of edit tracking
}
\description{
Which caching method should be used based on how edits are tracked
}
\details{
Some layers don't support edit tracing and can't be cached. This function checks whether it is supported.
}
