% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_esri_error}
\alias{check_esri_error}
\title{Check ESRI error}
\usage{
check_esri_error(content)
}
\arguments{
\item{content}{the content of a response}
}
\value{
null
}
\description{
Checks for an esri error and returns it
}
