% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{clear_layer_cache}
\alias{clear_layer_cache}
\title{Clear Layer Cache}
\usage{
clear_layer_cache(endpoint)
}
\arguments{
\item{endpoint}{the endpoint for the layer to remove the cache}
}
\description{
Clear the cache for a specific endpoint
}
\details{
This function deletes cache files which relate to an endpoint
}
