% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_url_parameters.R
\name{collapse_url_parameters}
\alias{collapse_url_parameters}
\title{Collapse URL Parameters}
\usage{
collapse_url_parameters(x, drop_null = FALSE)
}
\arguments{
\item{x}{a vector or list of url components}

\item{drop_null}{logical value should null list elements be dropped}
}
\value{
a string of query parameters
}
\description{
Helper function to collapse parameters into a url
}
