% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{combine_parse_esri_json}
\alias{combine_parse_esri_json}
\title{Parse GeoJSON list}
\usage{
combine_parse_esri_json(json_list, has_geometry = TRUE)
}
\arguments{
\item{has_geometry}{does the json h}

\item{geojson_list}{a list of geojson strings to be parsed into an sf object}
}
\value{
an sf object of all geojson parts combined
}
\description{
Parse a list of geojson strings
}
\details{
This function parses a list of geojson strings. It is a faster method
than parsing geojson individually and then combining them into one dataframe so speeds
up the process of getting data. The improved speed is most noticable for large geojson strings
or where there are many parts. Performance improvements won't be noticable for small lists
}
