% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{construct_cache_path}
\alias{construct_cache_path}
\title{Construct Cache Path}
\usage{
construct_cache_path(endpoint, create_dir = TRUE)
}
\arguments{
\item{endpoint}{the endpoint or service url of the layer}

\item{cache_dir}{the caching directory to be used (usually project specific)}
}
\value{
a string defining the cache path
}
\description{
Construct a path based upon the service url
}
\details{
This function takes a service url and a cache directory and creates a path that is unique to the service url
}
