% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{convert_datetimes_to_iso8601}
\alias{convert_datetimes_to_iso8601}
\title{Convert Datetimes to ISO8601}
\usage{
convert_datetimes_to_iso8601(x)
}
\arguments{
\item{x}{a dataframe that may contain datetime fields to convert to iso 8601}
}
\value{
a dataframe with converted dttms
}
\description{
Convert datetime fields to the ISO8601 standard
}
