% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimes.R
\name{convert_tz}
\alias{convert_tz}
\title{Convert Time Zones}
\usage{
convert_tz(x, tz, dttm_fields)
}
\arguments{
\item{x}{a dataframe}

\item{tz}{the timezone to convert datetime columns to}

\item{dttm_fields}{a character vector of field names to convert to \code{tz}}
}
\description{
Convert the time zones of a layer to a specified TZ
}
\details{
This function converts each datetime column to a specified timezone. This is a
superficial conversion and doesn't actually change the moment of time, just the way it is
displayed

The function is neccessary because sf::st_read returns datetimes in a different TZ (GMT/BST)
to the API. So when returning cached data, you will get a different result to data from
the api.
}
