% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimes.R
\name{detect_dttm_fields}
\alias{detect_dttm_fields}
\title{Detect Datetime Fields}
\usage{
detect_dttm_fields(x, layer_details)
}
\arguments{
\item{x}{a dataframe from a query}

\item{layer_details}{the layer details object returned by \code{get_layer_details}}
}
\value{
a character vector of field names which are present in x and have the type Datetime
}
\description{
Detect datetime fields which are present in a dataframe
}
\details{
This function uses the layer_details object to determine which columns have the datetime type
and then only returns those which are present in the dataframe.
}
