% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_lookup.R
\name{domain_lookup}
\alias{domain_lookup}
\title{Domain lookup}
\usage{
domain_lookup(layer_details)
}
\arguments{
\item{layer_details}{a list object generated by \code{get_layer_details}}
}
\value{
a tibble with three columns: the field name, the descriptive value and the coded value
}
\description{
Get the domains for each coded field
}
\details{
This function accepts an object generated by \code{get_layer_details} and returns a look up table of the coded domains
}
