% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{edit_config}
\alias{edit_config}
\title{Edit Cofig}
\usage{
edit_config(changes)
}
\arguments{
\item{changes}{a 1 level named list of config to edit}
}
\description{
Edit the config file and update the config env variables
}
