% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{empty_cache}
\alias{empty_cache}
\title{Empty the cache}
\usage{
empty_cache(force = FALSE)
}
\arguments{
\item{force}{should the function bypass asking for user confirmation to proceed? default is to ask for confirmation (FALSE)}
}
\description{
Remove all caching files
}
\details{
This function removes all caching files used by getarc
}
