% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\docType{data}
\name{endpoints}
\alias{endpoints}
\title{Endpoints}
\format{
An object of class \code{list} of length 6.
}
\usage{
endpoints
}
\description{
Endpoints for querying
}
\details{
A named list of endpoints for open source data that can be used when writing documentation and testing the package.
}
\keyword{datasets}
