% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_query.R
\name{esri_geometry_type}
\alias{esri_geometry_type}
\title{ESRI Geometry Type}
\usage{
esri_geometry_type(x)
}
\arguments{
\item{x}{an sf or sfc object}
}
\value{
a character string detailing the geometry type
}
\description{
Convert SF to ESRI geometry types
}
\details{
This finds the corresponding esri types for an sf object.
}
