% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri_sf_type_lookup.R
\docType{data}
\name{esri_sf_type_lookup}
\alias{esri_sf_type_lookup}
\title{Esri Type Lookup}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 3 columns.
}
\usage{
esri_sf_type_lookup
}
\description{
Find the Esri Type for SF types
}
\details{
The data is a tibble of esri feature types and their corresponding sf feature types.
}
\keyword{datasets}
