% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_query.R
\name{esri_spatial_filter}
\alias{esri_spatial_filter}
\title{Esri Spatial Filter}
\usage{
esri_spatial_filter(spatial_filter)
}
\arguments{
\item{spatial_filter}{the short name for the spatial filter. Should be one of intersects, contains, crosses, envelope_intersects, index_intersects, overlaps, touches or within}
}
\value{
the full name of the esri filter
}
\description{
Filter Name to Esri
}
\details{
This function converts the shortened version of an esri filter to it's esri filter name
}
