% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_server_endpoint.R
\name{feature_server_endpoint}
\alias{feature_server_endpoint}
\title{Feature Server Endpoint}
\usage{
feature_server_endpoint(host, instance, feature_server, layer_id = 0)
}
\arguments{
\item{host}{a string, defining the url for the ArcGIS Server host e.g 'https://services.arcgis.com'}

\item{instance}{a string defining the ArcGIS Server instance}

\item{feature_server}{a string defining the name of the feature server}

\item{layer_id}{an integar defining the layer id (beginning at 0)}
}
\value{
a string defining the endpoint url
}
\description{
Create a feature server endpoint string from url parameters
}
