% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{field_names}
\alias{field_names}
\title{Field Names}
\usage{
field_names(layer_details)
}
\arguments{
\item{layer_details}{the layer_details object}
}
\value{
a character vector of field names
}
\description{
Get the field names from layer details
}
