% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_oauth2.0_error.R
\name{find_oauth2.0_error}
\alias{find_oauth2.0_error}
\title{Find oauth2.0 error}
\source{
https://cran.r-project.org/package=httr
}
\usage{
find_oauth2.0_error(response)
}
\arguments{
\item{response}{a respoonse object returned by a request to refresh a token}
}
\description{
Find oauth errors
}
\details{
This function is borrowed from the httr package
}
