% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{generate_token}
\alias{generate_token}
\title{Generate a token}
\usage{
generate_token(endpoint, username, password, expiration = 60)
}
\arguments{
\item{endpoint}{the endpoint against which to request the token. this is generally in the format of:
https://<host>:<port>/<site>/tokens/generateToken
e.g: https://sampleserver6.arcgisonline.com/arcgis/tokens/generateToken}

\item{username}{your ArcGIS online username. You should make an effort to protect this by accessing it from an
environment variable so that it isn't saved in your source code or .Rhistory file.}

\item{password}{your ArcGIS online password. You should make an effort to protect this by accessing it from an
environment variable so that it isn't saved in your source code or .Rhistory file.}

\item{expiration}{The token expiration time in minutes (defaults to 60)}
}
\description{
Generate ArcGIS Access tokens with credentials
}
\details{
This functino enables a user to generate a token for accessing
ArcGIS services where Oauth2.0 isn't possible. The documentation is here:
https://developers.arcgis.com/rest/services-reference/generate-token.htm
You should make an effort to protect your username in password so that it isn't saved in the source code.
Some good options are here:
https://cran.r-project.org/web/packages/httr/vignettes/secrets.html

You could write a wrapper function around this function which gets your credentials from an
environment variable, calls the function and returns a token. getarc may provide this functionality eventually.
}
