% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_by_fids.R
\name{get_by_fids}
\alias{get_by_fids}
\title{Get by FIDs}
\usage{
get_by_fids(
  endpoint,
  query,
  my_token,
  return_geometry,
  return_n,
  layer_details,
  out_fields,
  object_ids = NULL
)
}
\arguments{
\item{endpoint}{a string defining the enpoint url.
It can be generated by the \code{feature_server_endpoint} and \code{map_server_endpoint} functons.
See https://developers.arcgis.com/rest/services-reference/get-started-with-the-services-directory.htm}

\item{query}{the query to POST}

\item{my_token}{an access token acquired via \code{get_token}}

\item{return_geometry}{should the geometry be returned or just a table?}

\item{return_n}{how many features (maximum) should be returned by the query?}

\item{layer_details}{the layer details returned by the get_layer_details function}

\item{out_fields}{the fields of the layer to return (character vector)}

\item{object_ids}{a vector of object IDs to return if this argument is NULL the function will get them
but passing them in aids performance if they have already been returned}
}
\value{
a tibble or sf object
}
\description{
Get a layer by querying the FIDs
}
\details{
This function works by checking if the requested return count is less tha the max record count.
If so, it doesn't bother with getting the FIDs and just requests the data and returns it.
Getting FIDs is a big overhea so this should be avoided where possible.
}
