% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_id.R
\name{get_count}
\alias{get_count}
\title{Get Count}
\usage{
get_count(endpoint, query = NULL, my_token = NULL)
}
\arguments{
\item{endpoint}{the url endpoint of the service. See feature_server_endpoint and map_server_endpoint}

\item{query}{an optional query to include in the request}

\item{my_token}{an access token to be included when required}
}
\value{
the count of features mathing the query
}
\description{
This function accepts an endpoint and a query and returns the count of features matching the query. It is a useful and fast method of determining if the query will return any data.
}
