% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(query_url, query, pb = NULL, my_token)
}
\arguments{
\item{query_url}{the query url which is passed to httr::POST()}

\item{query}{the query to POST}

\item{pb}{progress bar - default is NULL for no progress bar}

\item{my_token}{the access token or function used to generate one}
}
\value{
either a tibble or sf object depending on return_geometry
}
\description{
Get data from an endpoint
}
\details{
This function accepts a query URL and extracts a tibble from the response.
When a returnGeometry = "false" query was used previously
get_geojson wouldn't parse the data correctly and would return an empty tibble
A new function get_tibble has been added to use a different method for requesting and parsing data
when the geometry isn't returned.
}
