% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_details.R
\name{get_feature_details}
\alias{get_feature_details}
\title{Get feature server details}
\usage{
get_feature_details(host, instance, feature_server, my_token = NULL)
}
\arguments{
\item{host}{url for the ArcGIS Server host e.g 'https://services9.arcgis.com'}

\item{instance}{a string defining the ArcGIS Server instance}

\item{feature_server}{a string defining the name of the feature server}

\item{my_token}{an access token acquired via \code{get_token}}
}
\description{
Download feature layer from a feature server
}
