% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_id.R
\name{get_feature_ids}
\alias{get_feature_ids}
\title{Get Feature IDs}
\usage{
get_feature_ids(endpoint, query = NULL, my_token = NULL)
}
\arguments{
\item{endpoint}{the url endpoint of the service. See feature_server_endpoint and map_server_endpoint}

\item{query}{an optional query to include in the request}

\item{my_token}{an access token to be included when required}
}
\value{
a vector of feature IDs
}
\description{
Get the feature IDs of a layer
}
\details{
This function accepts an endpoint and a query and returns the feature_ids for an enpoint
}
