% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_server_details.R
\name{get_feature_server_details}
\alias{get_feature_server_details}
\title{Get Server Details}
\usage{
get_feature_server_details(endpoint, my_token = NULL)
}
\arguments{
\item{endpoint}{the server endpoint}

\item{my_token}{an access token generated by get_token, if neccessary for service}
}
\value{
a list
}
\description{
Get the details of a feature or map server
}
