% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_geojson}
\alias{get_geojson}
\title{Get GeoJSON}
\usage{
get_geojson(query_url)
}
\arguments{
\item{query_url}{the query url which is passed to httr::GET()}
}
\value{
an sf object
}
\description{
Download geojson from a query url
}
\details{
This function downloads geojson from a query_url, writes it to a temporary file and reads it in usinf st_read
}
