% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer.R
\name{get_layer}
\alias{get_layer}
\title{Get feature data}
\usage{
get_layer(
  host,
  instance,
  feature_server_name,
  layer_name,
  my_token,
  crs = 4326,
  layer_id = 0
)
}
\arguments{
\item{host}{}
}
\description{
Download feature layer from a feature server
}
