% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer_details.R
\name{get_layer_details}
\alias{get_layer_details}
\title{Get Layer Details}
\usage{
get_layer_details(endpoint, my_token = NULL)
}
\arguments{
\item{endpoint}{the endpoint of the layer (can be generated by either \code{feature_server_endpoint} or \code{map_server_endpoint})}

\item{my_token}{an access token acquired by \code{get_token}}
}
\value{
a list of parsed json
}
\description{
Get the details of a layer
}
