% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_tibble}
\alias{get_tibble}
\title{Get Tibble}
\usage{
get_tibble(query_url, query)
}
\arguments{
\item{query_url}{the query url which is passed to httr::POST()}

\item{query}{the query to POST}
}
\value{
a tibble
}
\description{
Get a Tibble from an endpoint
}
\details{
This function accepts a query URL and extracts a tibble from the response.
}
