% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_unique_id_field}
\alias{get_unique_id_field}
\title{Get ID field}
\usage{
get_unique_id_field(endpoint, layer_details)
}
\arguments{
\item{endpoint}{the layer endpoint}

\item{layer_details}{the layer details object returned by get_layer_details.}
}
\value{
a string defining the field that encodes the unique IDs
}
\description{
Get the unique ID field from the layer details
}
\details{
MapServers and Feature Servers encode the unique id field differently. This function uses two methods to retrieve the ID field from the layer details. Unique IDs are used for caching.
}
