% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{id_query}
\alias{id_query}
\title{ObjectID Query}
\usage{
id_query(object_id_name, object_ids, map_server)
}
\arguments{
\item{object_id_name}{the field name for object ids}

\item{object_ids}{a numeric vector of object IDs}

\item{map_server}{TRUE/FALSE is the service a map server}
}
\value{
a list with one named character element
}
\description{
Generate an object ID query to get data
}
\details{
This function takes object IDs and properly formats a query parameter based on whether the service
is a map server or feature server. For Map Servers the query should be in the form objectIds=1,2,3
and for Feature Servers the query should be where=ObjectId IN ('1', '2', ...)
}
