% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_layer.R
\name{init_cache_object}
\alias{init_cache_object}
\title{Init Cache Object}
\usage{
init_cache_object(
  data_cache = NULL,
  object_ids = NULL,
  query = NULL,
  method = NULL,
  any_changes = TRUE,
  use_cache = FALSE,
  cache_path = NULL,
  id_field = NULL
)
}
\arguments{
\item{data_cache}{the cached data}

\item{object_ids}{the object ids to request}

\item{query}{the query to execute}

\item{method}{the caching method to use}

\item{any_changes}{Have there been any changes to the data since the last cache?}

\item{cache_path}{where is/should the data be saved}

\item{id_field}{which is the unique id field}
}
\value{
a cache_object
}
\description{
Initialise the cache object
}
