% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{is_table}
\alias{is_table}
\title{Detect Tables}
\usage{
is_table(layer_details)
}
\arguments{
\item{layer_details}{the layer details object returnd by get_layer_details}
}
\value{
TRUE/FALSE
}
\description{
Is the endpoint a table
}
\details{
This function finds out if the layer is a table from the layer_details
}
