% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimes.R
\name{layer_timezone}
\alias{layer_timezone}
\title{Timezone}
\usage{
layer_timezone(layer_details, default = "UTC")
}
\arguments{
\item{layer_details}{the layer details object}

\item{default}{the default timezone - UTC}
}
\value{
a string defining the timezone, either form the layer details or the default
}
\description{
Get the layers timezone for parsing
}
