% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config_to_env}
\alias{load_config_to_env}
\title{Load Config}
\usage{
load_config_to_env(config)
}
\arguments{
\item{config}{a named list of config (supplied by read_config())}
}
\description{
Load the config into Env variables
}
