% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{lower_logical}
\alias{lower_logical}
\title{Lower Logical}
\usage{
lower_logical(x)
}
\arguments{
\item{x}{logical vector}
}
\value{
charactor vector
}
\description{
Convert a logical value to a lower case string
}
\details{
This function converts TRUE or FALSE to the equivalent lower case string "true" or "false" for use in queries.
}
