% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_empty_tibble}
\alias{make_empty_tibble}
\title{Make Empty Tibble}
\usage{
make_empty_tibble(field_names, out_fields)
}
\arguments{
\item{field_names}{the names of the columns to include}

\item{out_fields}{the fields to filter from the field_names if user has specified wanting certain fields}
}
\value{
an emtpy tibble with the names in out_fields
}
\description{
Make an empty tibble with specified column names
}
\details{
This function makes an empty tibble with specified column names
}
