% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_query_function.R
\name{make_query_function}
\alias{make_query_function}
\title{Make a query function}
\usage{
make_query_function(
  server_type,
  host,
  instance,
  feature_server = NULL,
  folder = NULL,
  service_name = NULL,
  layer_id = 0
)
}
\arguments{
\item{server_type}{one of c("MapServer", "FeatureServer")}

\item{host}{url for the ArcGIS Server host e.g 'https://services9.arcgis.com'}

\item{instance}{a string defining the ArcGIS Server instance}

\item{feature_server}{a string defining the name of the feature server (only required for feature servers)}

\item{folder}{a string defining the data folder (only required for map servers)}

\item{service_name}{a string defining the service name (only required for map servers)}

\item{layer_id}{an integar defining the layer id (starting at 0)}
}
\value{
a function that can be used to query the features of that layer
}
\description{
Specify the request parameters of a layer to generate a query function
}
