% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{map_server}
\alias{map_server}
\title{Map Server}
\usage{
map_server(endpoint)
}
\arguments{
\item{endpoint}{a string defining an esri endpoint}
}
\value{
TRUE or FALSE
}
\description{
Is an endpoint a map server
}
\details{
Feature & Map servers have different formats of querying and data. This function checks an endpoint string and returns TRUE or FALSE.
}
