% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_server_endpoint.R
\name{map_server_endpoint}
\alias{map_server_endpoint}
\title{Map server endpoint}
\usage{
map_server_endpoint(
  host,
  instance = "arcgis",
  folder,
  service_name,
  layer_id = 0
)
}
\arguments{
\item{host}{url for the ArcGIS Server host e.g 'https://services.arcgis.com'}

\item{instance}{a string defining the ArcGIS Server instance. The default is "arcgis"}

\item{folder}{a string defining the folder name}

\item{service_name}{a string defining the name of the service name}

\item{layer_id}{an integar defining the layer id (starting at 0)}
}
\value{
a string defining the endpoint url
}
\description{
Generate Map Server Endpoint url
}
