% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{modifyList}
\alias{modifyList}
\title{Modify vector}
\usage{
utils::modifyList(x, y)
}
\arguments{
\item{x}{a named vector to replace}

\item{y}{a named vector to combine with x}
}
\value{
a named vector that is the combination of x and y, not containing any elements in x with matches in y
}
\description{
Modify a vector using names
}
\details{
This function combines two named vectors, replacing elements in x where their names have matches in y.
}
