% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_access_token.R
\name{parse_access_token}
\alias{parse_access_token}
\title{Parse Access Token}
\usage{
parse_access_token(my_token)
}
\arguments{
\item{my_token}{an access token acquired via get_token}
}
\value{
the access token as a string
}
\description{
Helper function to parse access tokens
}
