% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_coded_domains.R
\name{parse_coded_domains}
\alias{parse_coded_domains}
\title{Parse Coded Domains}
\usage{
parse_coded_domains(data, domains)
}
\arguments{
\item{data}{a dataframe / sf object returned by \code{get_goejson}}

\item{domains}{a dataframe of domains and the descriptive and coded values returned by \code{domain_lookup}}
}
\value{
an sf object with coded domain values exchanged for their coded values
}
\description{
Replace the coded domain values with their descriptive values
}
