% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimes.R
\name{parse_datetimes}
\alias{parse_datetimes}
\title{Parse Datetimes}
\usage{
parse_datetimes(data, layer_details)
}
\arguments{
\item{data}{a dataframe / sf object returned by \code{get_geojson}}

\item{layer_details}{a list object returned by \code{get_layer_details}}
}
\value{
a dataframe or sf object of the same structure as \code{data} but with datetime variables parsed
}
\description{
Parse Datetime variables
}
