% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_esri_json.R
\name{parse_esri_data}
\alias{parse_esri_data}
\title{Parse ESRI JSON}
\usage{
parse_esri_data(x, geometry)
}
\arguments{
\item{x}{an R list of data returned by the APPI (parsed JSON returned by the API)}

\item{geometry}{is there a geometry element to the json? TRUE/FALSE}
}
\value{
a tibble or sf object
}
\description{
Parse the JSON response of the API
}
\details{
Conditionally parse the JSON response of the API depending on whether it is a table of layer (e.g. has geometry)
}
