% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimes.R
\name{parse_esri_datetime}
\alias{parse_esri_datetime}
\title{Parse esri datetime}
\usage{
parse_esri_datetime(x, tz)
}
\arguments{
\item{x}{a vector of datetimes in the esri format}

\item{tz}{the timezone to return}
}
\value{
a vector of datetimes
}
\description{
Parse datetimes stored by esri
}
