% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_esri_json.R
\name{parse_esri_json}
\alias{parse_esri_json}
\title{Parse ESRI JSON}
\usage{
parse_esri_json(json_list, geometry)
}
\arguments{
\item{json_list}{an R list of JSON returned by the API (in the format of ESRI JSON, not GEOJSON)}

\item{geometry}{is there a geometry element to the json? TRUE/FALSE}
}
\value{
a tibble or sf object
}
\description{
Parse the JSON response of the API
}
\details{
Conditionally parse the JSON response of the API depending on whether it is a table of layer (e.g. has geometry)
}
