% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{parse_esri_json_table}
\alias{parse_esri_json_table}
\title{Parse ESRI JSON tables}
\usage{
parse_esri_json_table(json)
}
\arguments{
\item{json}{ESRI json representing a table}
}
\description{
Parse ESRI JSON (without geometry) into a table
}
\details{
This function accepts esrijson and parses it as a table
}
