% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimes.R
\name{parse_iso8601_datetime}
\alias{parse_iso8601_datetime}
\title{Parse ISO801 datetimes}
\usage{
parse_iso8601_datetime(x)
}
\arguments{
\item{x}{a character vector to parse fitting iso8601 standard}
}
\value{
a POSIXct vector
}
\description{
Parse datetimes fitting iso standard
This function takes a character vector of iso datetimes & parses them to a posixct class
It checks that it isn't already posixct
}
