% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_esri_json.R
\name{parse_json_layer}
\alias{parse_json_layer}
\title{Parse JSON Layer}
\usage{
parse_json_layer(json)
}
\arguments{
\item{json}{ESRI JSON (not GeoJSON)}
}
\value{
an sf object
}
\description{
Parse an ESRI json response into an SF object
}
