% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_esri_json.R
\name{parse_json_table}
\alias{parse_json_table}
\title{Parse JSON Table}
\usage{
parse_json_table(json)
}
\arguments{
\item{json}{ESRI JSON (not GeoJSON)}
}
\value{
a tibble
}
\description{
Parse an ESRI json response into a table
}
