% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_esri_json.R
\name{parse_layer}
\alias{parse_layer}
\title{Parse JSON Layer}
\usage{
parse_layer(x)
}
\arguments{
\item{x}{ESRI JSON (not GeoJSON)}
}
\value{
an sf object
}
\description{
Parse an ESRI json response into an SF object
}
