% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_query_url.R
\name{parse_query_url}
\alias{parse_query_url}
\title{Parse Query URL}
\usage{
parse_query_url(url)
}
\arguments{
\item{url}{a string defining the url of the query.}
}
\value{
a function created by \code{specifiy_layer_params}
}
\description{
Parse a query url to return an R function to get the data
}
\details{
This function enables the user to supply a
query url for a data source and it will parse the parameters returning a function
}
\examples{
# Specify the URL for SSSIs
url <- "https://services.arcgis.com/JJzESW51TqeY9uat/arcgis/rest/services/SSSI_England/FeatureServer/0/query?where=1\%3D1&outFields=*&outSR=4326&f=json"
# Parse the url into a get function
get_sssi <- parse_query_url(url)
# Get the data, passing an additional parameter to only return one feature
one_sssi <- get_sssi(query = c("resultRecordCount" = "1"))
}
