% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{parse_rjson}
\alias{parse_rjson}
\title{Parse R JSON}
\usage{
parse_rjson(response)
}
\arguments{
\item{response}{a response object returned by a GET or POST request}
}
\value{
an R object of parsed json
}
\description{
Parse a httr response using rjson
}
\details{
This function implements faster json parsing the jsonlite and httr
}
