% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{parse_types}
\alias{parse_types}
\title{Parse Types}
\usage{
parse_types(x, layer_details)
}
\arguments{
\item{x}{the dataframe to parse}

\item{layer_details}{the layer details object}
}
\value{
a dataframe like x but with variabes matching the specified types
}
\description{
Parse column types
}
\details{
Parse the tibbles column types using the layer details obect to define them
}
