% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_layer.R
\name{query_layer}
\alias{query_layer}
\title{Query Layer}
\usage{
query_layer(
  endpoint,
  query = NULL,
  bounding_box = NULL,
  crs = 4326,
  my_token = NULL,
  return_geometry = TRUE
)
}
\arguments{
\item{endpoint}{a string defining the enpoint url.
It can be generated by the \code{feature_server_endpoint} and \code{map_server_endpoint} functons.
See https://developers.arcgis.com/rest/services-reference/get-started-with-the-services-directory.htm}

\item{query}{a named vector of parameters to include in the query}

\item{bounding_box}{an optional bounding box (generated by sf::st_bbox()) to perform a spatial intersects}

\item{crs}{the output crs, defaulting to 4326}

\item{my_token}{an access token acquired via \code{get_token}}

\item{return_geometry}{should the geom}
}
\value{
an sf object
}
\description{
Query a layer on an arcgis server
}
