% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_layer_fs.R
\name{query_layer_fs}
\alias{query_layer_fs}
\title{Get Feature Server layer}
\usage{
query_layer_fs(
  host,
  instance,
  feature_server,
  layer_id = 0,
  query = NULL,
  bounding_box = NULL,
  crs = 4326,
  my_token = NULL
)
}
\arguments{
\item{host}{url for the ArcGIS Server host e.g 'https://services9.arcgis.com'}

\item{instance}{a string defining the ArcGIS Server instance}

\item{feature_server}{a string defining the name of the feature server}

\item{layer_id}{an integar defining the layer id (starting at 0)}

\item{query}{a named vector of parameters to include in the query}

\item{bounding_box}{an optional bounding box (generated by sf::st_bbox()) to perform a spatial intersects}

\item{crs}{the output crs, defaulting to 4326}

\item{my_token}{an access token acquired via \code{get_token}}
}
\value{
an sf object
}
\description{
Download feature layer from a feature server
}
