% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_layer_ms.R
\name{query_layer_ms}
\alias{query_layer_ms}
\title{Query Map Server Layer}
\usage{
query_layer_ms(
  host,
  instance = "arcgis",
  folder,
  service_name,
  layer_id = 0,
  crs = 4326,
  bounding_box = NULL,
  my_token = NULL,
  query = NULL
)
}
\arguments{
\item{host}{url for the ArcGIS Server host e.g 'https://services.arcgis.com'}

\item{instance}{a string defining the ArcGIS Server instance. The default is "arcgis"}

\item{folder}{a string defining the folder name}

\item{service_name}{a string defining the name of the service name}

\item{layer_id}{an integar defining the layer id (starting at 0)}

\item{crs}{the output crs, defaulting to 4326}

\item{bounding_box}{an optional bounding box (generated by sf::st_bbox()) to perform a spatial intersects}

\item{my_token}{an access token acquired via \code{get_token}}

\item{query}{a named vector of parameters to include in the query}
}
\value{
an sf object
}
\description{
Download feature layer from a Map Server
}
