% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{query_object}
\alias{query_object}
\title{Query Object}
\usage{
query_object(
  default = default_query_parameters(),
  user_query = list(),
  mandatory = list(),
  token = NULL
)
}
\arguments{
\item{default}{the default query parameters such as return format and crs (list).}

\item{user_query}{the user query passed into the query argument or as argument params (list).}

\item{mandatory}{the mandatory query parameters (list).}

\item{token}{the access token if required.}
}
\value{
a list of query parameters to pass as the post body in a request
}
\description{
Compose a Query Object
}
\details{
This function composes a query object from a set of query parameters that are either default, mandatory or user specified.
}
