% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_layer.R
\name{refresh_cache}
\alias{refresh_cache}
\title{Refresh cache}
\usage{
refresh_cache(data, cache_object)
}
\arguments{
\item{data}{the newly downloaded data}

\item{cache_object}{the cache object generated by init_cache}
}
\value{
a tibble
}
\description{
Refresh the data cache
}
