% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_token.R
\name{refresh_token}
\alias{refresh_token}
\title{Refresh Token}
\source{
https://rdrr.io/cran/httr/src/R/oauth-refresh.R https://cran.r-project.org/package=httr
}
\usage{
refresh_token(
  endpoint,
  app,
  credentials,
  user_params = NULL,
  use_basic_auth = FALSE
)
}
\arguments{
\item{endpoint}{the endpoint to authenticae against}

\item{app}{the app to autheticate}

\item{credentials}{the credentials list from a token generated by get_token}

\item{user_params}{user parameters to send with request}

\item{use_basic_auth}{should basic authorization be used - don't know what this does}
}
\value{
a list of refreshed credentials
}
\description{
Refresh  Oauth 2.0 Token
}
\details{
This function is baseed upon httr:::refresh_oauth2.0 which doesn't parse the credentials into a list.
An additional call to rjson::fromJSON has been added to parse the credentials correctly
}
