% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{replace_null}
\alias{replace_null}
\title{Replace NULLS}
\usage{
replace_null(x, with = NA)
}
\arguments{
\item{x}{a vector possibly containing nulls}

\item{with}{what to replace the nulls with, default is NA}
}
\value{
a vector of equal length to x which will not contain null values but the value of with instead
}
\description{
Replace NULL values with a supplied value
}
\details{
This function is a helper to replace null values returned by the API with a specified values. This is NA by default.
}
