% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{set_cache_directory}
\alias{set_cache_directory}
\title{Set Cache Directory}
\usage{
set_cache_directory(directory, create = FALSE)
}
\arguments{
\item{directory}{the directory to use when caching}
}
\description{
Set the directory to be used when caching layers
}
