% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{set_credentials}
\alias{set_credentials}
\title{Get/Set Credentials}
\usage{
set_credentials(client_id, client_secret, app_name)
}
\description{
Get or set app credentials
}
\details{
set_credentials and get_credentials set and retrieve environment variables for the client id, client secret and app name.
This then allows get_token to retrieve the variables.
}
