% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sf_geojson_write}
\alias{sf_geojson_write}
\title{Write sf as GEOJSON}
\usage{
sf_geojson_write(sf, filepath)
}
\arguments{
\item{sf}{an sf object to be written to disk}

\item{filepath}{the filepath of the sf object}
}
\description{
Write a GeoJSON file
}
\details{
This function takes an sf object and writes it to a specified filepath as geojson
This is a faster alternative to sf::st_write
}
