% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_query.R
\name{sf_to_json}
\alias{sf_to_json}
\title{sf to json}
\usage{
sf_to_json(x)
}
\arguments{
\item{x}{an sf or sfc object}
}
\value{
a character string of json.
}
\description{
Convert an sfc object to json
}
\details{
This function accepts an sf object and converts it to json (not geojson). This format is required by esri for spatial queries.
}
