% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_query.R
\name{simplify_sf_to_nchar}
\alias{simplify_sf_to_nchar}
\title{Simplify sf to length}
\usage{
simplify_sf_to_nchar(x, char)
}
\arguments{
\item{x}{an sf object to simplify}

\item{char}{numeric - the number of characters to simplify to based on st_as_text}
}
\description{
Simplify SF to N Char
}
