% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_query.R
\name{spatial_query}
\alias{spatial_query}
\title{Spatial Query}
\usage{
spatial_query(x, spatial_filter = "esriSpatialRelIntersects")
}
\arguments{
\item{x}{an sf or sfc object}

\item{spatial_filter}{the spatial relationship of the filter to specify in the query. Default is esriSpatialRelIntersects.
Options are: esriSpatialRelIntersects, esriSpatialRelContains, esriSpatialRelCrosses, esriSpatialRelEnvelopeIntersects,
esriSpatialRelIndexIntersects, esriSpatialRelOverlaps, esriSpatialRelTouches or esriSpatialRelWithin.}
}
\value{
a list with geometry, geometryType, SpatialRel and inSR
}
\description{
Convert SF/SFC to spatial query.
}
\details{
This function accepts an sf/sfc object and converts it to a spatial query fitting the esri api spec.
}
