% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_query_to_list.R
\name{spatial_query_to_list}
\alias{spatial_query_to_list}
\title{Generate a Spatial Query}
\usage{
spatial_query_to_list(bbox)
}
\arguments{
\item{bbox}{a bounding box created by st_bbox()}
}
\value{
a list of spatial query elements
}
\description{
Generate a list of spatial query parameters from a bounding box
}
\details{
Deprecated, please use spatial_query instead
}
