% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_layer_params.R
\name{specify_layer_params}
\alias{specify_layer_params}
\title{Specify Layer Parameters}
\usage{
specify_layer_params(
  host,
  instance,
  feature_server,
  layer_id = 0,
  query = NULL
)
}
\arguments{
\item{host}{url for the ArcGIS Server host e.g 'https://services9.arcgis.com'}

\item{instance}{a string defining the ArcGIS Server instance}

\item{feature_server}{a string defining the name of the feature server}

\item{layer_id}{an integar defining the layer id (starting at 0)}
}
\value{
a function that can be used to query the features of that layer
}
\description{
Specify the request parameters of a layer
}
