% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{split_vector}
\alias{split_vector}
\title{Split a Vector}
\usage{
split_vector(x, max_length)
}
\arguments{
\item{x}{a vector}

\item{max_length}{the maximum length of the returned vectors}
}
\value{
a list of split x where each element does not exceed the length of max_length
}
\description{
Split a Vector
}
\details{
This function splits a vector so that it doesn't exceed a maximum length
}
