% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{supports_edit_tracking}
\alias{supports_edit_tracking}
\title{Supports Edit Tracking?}
\usage{
supports_edit_tracking(layer_details)
}
\arguments{
\item{layer_details}{the layer details object returned be get_layer_details.}
}
\value{
TRUE/FALSE. TRUE if the layer supports edit tracking
}
\description{
Does the layer support edit tracking
}
\details{
Some layers don't support edit tracing and can't be cached. This function checks whether it is supported.
}
