% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{token_expired}
\alias{token_expired}
\title{Token Expired}
\usage{
token_expired(my_token)
}
\arguments{
\item{my_token}{an access token generated by get_token}
}
\value{
logical value detailing whether the token has or has not expired
}
\description{
Has the Oauth token expired
}
