% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_functions.R
\docType{data}
\name{type_functions}
\alias{type_functions}
\title{Type Functions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 2 columns.
}
\usage{
type_functions
}
\description{
The function used to assert a variable matches the type returned by ESRI
}
\details{
st_read doesn't always return the right data type (e.g. if the variable is all NA) and
this causes issues when combining data. This table is used in the parse_types function
to parse each variable in a data frame to the correst type
}
\keyword{datasets}
