% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{where_in_query}
\alias{where_in_query}
\title{Where In Query}
\usage{
where_in_query(field, matching, named = FALSE)
}
\arguments{
\item{field}{which field should match \code{matching}}

\item{matching}{which elements of \code{field} should be returned}
}
\value{
a named character vector of length 1 to be included in query
}
\description{
Construct a where in query
}
\details{
Construct a where in query to se for getting FIDs
}
