% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_query_parameters.R
\name{collapse_query_parameters}
\alias{collapse_query_parameters}
\title{Collapse Query Parameters}
\usage{
collapse_query_parameters(x, drop_null = FALSE)
}
\arguments{
\item{x}{a named list of query parameters}

\item{drop_null}{logical value should null list elements be dropped}
}
\value{
a string of query parameters
}
\description{
Helper function to collapse parameters into a typical url query
}
