% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer.R
\name{get_layer}
\alias{get_layer}
\title{Get feature data}
\usage{
get_layer(
  host,
  instance,
  feature_server,
  my_token = NULL,
  crs = 4326,
  layer_id = 0,
  query = NULL
)
}
\arguments{
\item{host}{url for the ArcGIS Server host e.g 'https://services9.arcgis.com'}

\item{instance}{a string defining the ArcGIS Server instance}

\item{feature_server}{a string defining the name of the feature server}

\item{my_token}{an access token acquired via \code{get_token}}

\item{crs}{the output crs, defaulting to 4326}

\item{layer_id}{an integar defining the layer id (starting at 0)}

\item{query}{a named vector of parameters to include in the query}
}
\value{
an sf object
}
\description{
Download feature layer from a feature server
}
