% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_layer.R
\name{init_cache}
\alias{init_cache}
\title{Init Cache}
\usage{
init_cache(endpoint, query, cache, layer_details, my_token = NULL, id_field)
}
\arguments{
\item{endpoint}{the endpoint against which to query}

\item{query}{the query to execute}

\item{cache}{the cache path supplied by the user, can be null}

\item{layer_details}{the layer details argument supplied by get layer details}

\item{my_token}{the authentication token returned by get_token or other oauth method}

\item{id_field}{the id field against which to renew the cache}
}
\value{
a cache object defining the cached data & various bits of info to determine how the cache should be updated if at all
}
\description{
Initialise the cache
}
