% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_esri_datetime.R
\name{parse_esri_datetime}
\alias{parse_esri_datetime}
\title{Parse esri datetime}
\usage{
parse_esri_datetime(x)
}
\arguments{
\item{x}{a vector of datetimes in the esri format}
}
\value{
a vector of datetimes
}
\description{
Parse datetimes stored by esri
}
