% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_layer.R
\name{query_layer}
\alias{query_layer}
\title{Query Layer}
\usage{
query_layer(
  endpoint,
  in_geometry = NULL,
  spatial_filter = "intersects",
  return_geometry = TRUE,
  out_fields = c("*"),
  return_n = NULL,
  geometry_precision = NULL,
  query = NULL,
  crs = 4326,
  my_token = NULL,
  bounding_box = lifecycle::deprecated()
)
}
\arguments{
\item{endpoint}{a string defining the enpoint url.
It can be generated by the \code{feature_server_endpoint} and \code{map_server_endpoint} functons.
See https://developers.arcgis.com/rest/services-reference/get-started-with-the-services-directory.htm}

\item{in_geometry}{the geometry to use when applying a spatial filter to the layer. This should be an sf object (bounding boxes are accepted and are the fastest).}

\item{spatial_filter}{The type of spatial filter. Defaults to intersection (intersects).
Options are: intersects, contains, crosses, envelope_intersects, index_intersects, overlaps, touches, within}

\item{return_geometry}{should the geometry be returned or just a table?}

\item{out_fields}{the fields of the layer to return (character vector)}

\item{return_n}{how many features (maximum) should be returned by the query?}

\item{geometry_precision}{The number of decimal places in the response geometries returned by the query operation}

\item{query}{a named vector of parameters to include in the query}

\item{crs}{the output crs, defaulting to 4326}

\item{my_token}{an access token acquired via \code{get_token}}

\item{bounding_box}{DEPRECTED. An optional bounding box (generated by sf::st_bbox()) to perform a spatial intersects}
}
\value{
an sf object
}
\description{
Query a layer on an arcgis server
}
